refactor: consolidate websoc logic
